Optimize OS and Go version matrix strategy for CI workflow
before the CI workflow used a matrix strategy to run the `lint-node`,
`lint-go` and `test` jobs, but this also included steps that were not
necessary for this repository. This has been improved to make the
workflow run faster by avoiding unnecessary tasks:

- The `lint-node` job has been changed to only run on the currently
  latest stable Node version `14.x` [1] only on Linux because this
  repository is not focused on JavaScript but only runs Node based tools
  to lint other files within this repository.
- The `lint-go` job has been changed to only run on the currently latest
  stable Go version `1.15.x` [2] only on Linux because `golangci-lint`
  doesn't care about the Go version and OS it runs on but only
  statically checks the source code.
- The `test` job has been changed to only run on the currently latest
  stable Go version `1.15.x` [2] and only Linux and Windows while macOS
  is not necessary for this repository because there is no macOS
  specific code.

These changes also help to keep the required GitHub Action run minutes
for the account of this repository as small as possible without wasting
resources for unnecessary tasks.

[1]: https://github.com/nodejs/node/blob/master/doc/changelogs/CHANGELOG_V14.md#14.12.0
[2]: https://golang.org/doc/go1.15

GH-12