Split icon and prefix features into new `pencil` package (#19)
Before the `inkpen` package provided a writer that was able to print
verbosity level icons and prefixes with colors.
To allow to use icons and prefixes without colors both features have
been moved into the new `pencil` package with the new `pencil.Pencil`
type. The `inkpen.Inkpen` type has been changed to compose this new type
and additionally allow to color the output.

In order to make it easier to configure both types, new package specific
options have been implemented that are available via variadic parameters
to the constructor functions of these types. This reduces the required
code to configure a writer and also provides a safer usage with
Goroutines.

Resolves GH-18